requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.3
cycler==0.10.0
discord.py==1.7.3
Django==3.2.5
future==0.18.2
google==3.0.0
google-api-core==1.31.0
google-api-python-client==2.14.0
google-auth==1.33.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-spreadsheet==0.0.6
googleapis-common-protos==1.53.0
gspread==3.7.0
h5py==3.3.0
httplib2==0.19.1
idna==3.2
joblib==1.0.1
Keras==2.4.3
kiwisolver==1.3.1
matplotlib==3.4.2
multidict==5.1.0
numpy==1.21.0
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.0
pandas==1.3.0
pefile==2021.5.24
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pygame==2.0.1
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
PyNaCl==1.4.0
pyparsing==2.4.7
PySimpleGUI==4.46.0
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.1
threadpoolctl==2.2.0
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
yarl==1.6.3
